Starting Layout
I initialized a new project for the AP CS Prep App. It includes a navigation view scroll bar on the left, a settings button, and a compose message button(delete in the future). I also modifired the title of the text in nav view to something more suitable. Good start!