Basic Layout finish
I sucessfully created the base layout for the unit activity and create a method to handle all xml interactions and filling.
There is a listview of questions which are not currently mapped to anything yet.  In addition there is a new Unit1 textfile which can be used for testing how layout works on different units