Answer Selection Logic
I implemented a toString method for the question object so that we can easily debug and see what a question entails. In additon I created the baisc layout for the question page and it's looking pretty bad. It currently contains 6 TextViews(Question + 5 Answers) and a submission button. I currently am working on getting the answer selection logic to work as well. I'm thinking of having the answer's textview change borders when selected and the user can press submit after selection to file in their answer.