YESS
In this grind out learning xml session I finally learned/discovered how to put a background on textviews in both xml and java(for versatility). I implemented a file to show/demonstrate how to have a universal background be applied on textviews such as answers. In the end I was able to create a layout, using borders, that would visually show the currently selected answer.